Updated tests for TaskController

+++ src/main/java/rest/task/TaskService.java
@@ -38,11 +38,11 @@ public class TaskService {
 
     public Task updateTask(int id){
         Task task = taskRepository.findOne(id);
-        Task newTask = Task.copyTask(task);
+        if(task==null) return null;
 
+        Task newTask = Task.copyTask(task);
         if(task.isDone()) newTask.setDone(false);
         else newTask.setDone(true);
-
         taskRepository.delete(id);
         return taskRepository.save(newTask);
     }
